Ability to change configuration values via envvar
Not the best way to do it, since its unaware of scope, but it will do
for now.

Entrypoint script will now search for environment variables prefixed by
`TES3MP_SERVER_` and update their value in the matching variable in
`tes3mp-server-default.cfg`

Example:

podman run --rm -it \
    --name TES3MP-server \
    -e TES3MP_SERVER_HOSTNAME="C0da Enjoyers Server" \
    -e TES3MP_SERVER_MAXIMUM_PLAYERS=1 \
    -e TES3MP_SERVER_PASSWORD="superExclusive" \
    -p "25565:25565/udp" \
    tes3mp/server

Will change the value of `General.hostname` to `Test` and the value of
`General.maximumPlayers` to `1` (note the case change).